+++ tests/snapshot_tests/test_snapshots.py
@@ -162,6 +162,11 @@ def test_offsets(snap_compare):
     assert snap_compare("snapshot_apps/offsets.py")
 
 
+def test_nested_auto_heights(snap_compare):
+    """Test refreshing widget within a auto sized container"""
+    assert snap_compare("snapshot_apps/nested_auto_heights.py", press=["1", "2", "_"])
+
+
 # --- Other ---